Play logic closes #11

+++ src/main/java/ch/zhaw/guess/model/QuestionDTO.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @Setter
 public class QuestionDTO {
     private String questionText;
-    private double correctAnswer;
+    private int correctAnswer;
 }
